polyp_seg/requirements.txt
absl-py==2.3.1
certifi==2026.2.25
cffi==1.17.1
charset-normalizer==3.4.4
contourpy==1.1.1
cryptography==46.0.5
cycler==0.12.1
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
google-auth==2.48.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
hf-xet==1.3.2
huggingface_hub==0.36.2
idna==3.11
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
numpy==1.24.4
oauthlib==3.3.1
opencv-python==4.13.0.92
packaging==26.0
pillow==10.4.0
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
scipy==1.10.1
six==1.17.0
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
timm==0.4.12
torch==1.7.1+cu110
torchvision==0.8.2+cu110
tqdm==4.67.3
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
